Switching to Jekyll branch redcarpet-2.0
Redcarpet 2 is helpful because it included

``` text
  fenced_code_blocks
```

which are used by github (gollum, etc.).  This assumes that a working
version of Pygments has been installed.

Code blocks appear to be rendering fine, but upon examining the source,
span-enabled syntax highlighting does not appear, just PRE and
CODE CLASS = TEXT tags.